chore: abstracting callback functions to reduce the complexity of the jsx
