+++ libs/components/editor-core/src/block-pendant/pendant-operation-panel/UpdatePendantPanel.tsx
@@ -1,23 +1,13 @@
 import { useState } from 'react';
+import { Input, Tooltip } from '@toeverything/components/ui';
+import { HelpCenterIcon } from '@toeverything/components/icons';
 import { PendantModifyPanel } from '../pendant-modify-panel';
 import type { AsyncBlock } from '../../editor';
 import {
-    genSelectOptionId,
-    InformationProperty,
-    type MultiSelectProperty,
     type RecastBlockValue,
     type RecastMetaProperty,
-    type SelectOption,
-    type SelectProperty,
-    useRecastBlockMeta,
-    useSelectProperty,
 } from '../../recast-block';
-import { OptionType, PendantTypes, TempInformationType } from '../types';
-import {
-    getOfficialSelected,
-    getPendantConfigByType,
-    checkPendantForm,
-} from '../utils';
+import { getPendantConfigByType } from '../utils';
 import { usePendant } from '../use-pendant';
 import {
     StyledPopoverWrapper,
@@ -28,10 +18,8 @@ import {
     StyledPopoverSubTitle,
 } from '../StyledComponent';
 import { IconMap, pendantOptions } from '../config';
-import { Input, message, Tooltip } from '@toeverything/components/ui';
-import { HelpCenterIcon } from '@toeverything/components/icons';
 
-type SelectPropertyType = MultiSelectProperty | SelectProperty;
+import { useOnUpdateSure } from './hooks';
 
 type Props = {
     value: RecastBlockValue;
@@ -52,13 +40,12 @@ export const UpdatePendantPanel = ({
     onCancel,
     titleEditable = false,
 }: Props) => {
-    const { updateSelect } = useSelectProperty();
-    const { setPendant, removePendant } = usePendant(block);
     const pendantOption = pendantOptions.find(v => v.type === property.type);
     const iconConfig = getPendantConfigByType(property.type);
+    const { removePendant } = usePendant(block);
     const Icon = IconMap[iconConfig.iconName];
-    const { updateProperty } = useRecastBlockMeta();
-    const [fieldTitle, setFieldTitle] = useState(property.name);
+    const [fieldName, setFieldName] = useState(property.name);
+    const onUpdateSure = useOnUpdateSure({ block, property });
 
     return (
         <StyledPopoverWrapper>
@@ -76,10 +63,10 @@ export const UpdatePendantPanel = ({
             <StyledOperationLabel>Field Title</StyledOperationLabel>
             {titleEditable ? (
                 <Input
-                    value={fieldTitle}
+                    value={fieldName}
                     placeholder="Input your field name here"
                     onChange={e => {
-                        setFieldTitle(e.target.value);
+                        setFieldName(e.target.value);
                     }}
                     endAdornment={
                         <Tooltip content="Help info here">
@@ -110,126 +97,12 @@ export const UpdatePendantPanel = ({
                 property={property}
                 type={property.type}
                 onSure={async (type, newPropertyItem, newValue) => {
-                    const checkResult = checkPendantForm(
+                    await onUpdateSure({
                         type,
-                        fieldTitle,
                         newPropertyItem,
-                        newValue
-                    );
-
-                    if (!checkResult.passed) {
-                        await message.error(checkResult.message);
-                        return;
-                    }
-
-                    if (
-                        type === PendantTypes.MultiSelect ||
-                        type === PendantTypes.Select ||
-                        type === PendantTypes.Status
-                    ) {
-                        const newOptions = newPropertyItem as OptionType[];
-                        let selectProperty = property as SelectPropertyType;
-                        const deleteOptionIds = selectProperty.options
-                            .filter(o => {
-                                return !newOptions.find(no => no.id === o.id);
-                            })
-                            .map(o => o.id);
-                        const addOptions = newOptions.filter(
-                            o => typeof o.id === 'number'
-                        );
-
-                        const { addSelectOptions, removeSelectOptions } =
-                            updateSelect(selectProperty);
-
-                        deleteOptionIds.length &&
-                            (selectProperty = (await removeSelectOptions(
-                                ...deleteOptionIds
-                            )) as SelectPropertyType);
-
-                        addOptions.length &&
-                            (selectProperty = (await addSelectOptions(
-                                ...(addOptions as unknown as Omit<
-                                    SelectOption,
-                                    'id'
-                                >[])
-                            )) as SelectPropertyType);
-
-                        const selectedId = getOfficialSelected({
-                            isMulti: type === PendantTypes.MultiSelect,
-                            options: selectProperty.options,
-                            tempOptions: newPropertyItem,
-                            tempSelectedId: newValue,
-                        });
-
-                        await setPendant(selectProperty, selectedId);
-                    } else if (type === PendantTypes.Information) {
-                        // const { emailOptions, phoneOptions, locationOptions } =
-                        //     property as InformationProperty;
-                        const optionGroup =
-                            newPropertyItem as TempInformationType;
-
-                        const emailOptions = optionGroup.emailOptions.map(
-                            option => {
-                                if (typeof option.id === 'number') {
-                                    option.id = genSelectOptionId();
-                                }
-                                return option;
-                            }
-                        );
-                        const phoneOptions = optionGroup.phoneOptions.map(
-                            option => {
-                                if (typeof option.id === 'number') {
-                                    option.id = genSelectOptionId();
-                                }
-                                return option;
-                            }
-                        );
-                        const locationOptions = optionGroup.locationOptions.map(
-                            option => {
-                                if (typeof option.id === 'number') {
-                                    option.id = genSelectOptionId();
-                                }
-                                return option;
-                            }
-                        );
-
-                        const newProperty = await updateProperty({
-                            ...property,
-                            emailOptions,
-                            phoneOptions,
-                            locationOptions,
-                        } as InformationProperty);
-
-                        await setPendant(newProperty, {
-                            email: getOfficialSelected({
-                                isMulti: true,
-                                options: emailOptions as SelectOption[],
-                                tempOptions: newPropertyItem.emailOptions,
-                                tempSelectedId: newValue.email,
-                            }),
-                            phone: getOfficialSelected({
-                                isMulti: true,
-                                options: phoneOptions as SelectOption[],
-                                tempOptions: newPropertyItem.phoneOptions,
-                                tempSelectedId: newValue.phone,
-                            }),
-                            location: getOfficialSelected({
-                                isMulti: true,
-                                options: locationOptions as SelectOption[],
-                                tempOptions: newPropertyItem.locationOptions,
-                                tempSelectedId: newValue.location,
-                            }),
-                        });
-                    } else {
-                        await setPendant(property, newValue);
-                    }
-
-                    if (fieldTitle !== property.name) {
-                        await updateProperty({
-                            ...property,
-                            name: fieldTitle,
-                        });
-                    }
+                        newValue,
+                        fieldName,
+                    });
                     onSure?.();
                 }}
                 onDelete={
